+++ packages/deploy-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/deploy-config-sub-generator
 
+## 0.0.5
+
+### Patch Changes
+
+-   625fb21: ensure destination is appended to backend config
+
+## 0.0.4
+
+### Patch Changes
+
+-   026dfe0: fix undefined error when no backend config
+
 ## 0.0.3
 
 ### Patch Changes